requirements.txt
python-dotenv
azure-search-documents=11.6.0b7
azure-identity
azure-core
openai
prompty[azure]
prompty[tracer]
promptflow-azure